+++ pmd-core/src/main/java/net/sourceforge/pmd/util/designer/CodeEditorTextPane.java
@@ -8,7 +8,7 @@ import javax.swing.JTextPane;
 
 import net.sourceforge.pmd.lang.ast.Node;
 
-@Deprecated
+@Deprecated // to be removed with PMD 7.0.0
 public class CodeEditorTextPane extends JTextPane implements LineGetter {
 
     private String[] getLines() {