+++ pmd-core/src/main/java/net/sourceforge/pmd/util/designer/CreateXMLRulePanel.java
@@ -24,7 +24,7 @@ import net.sourceforge.pmd.PMD;
  * This class is responsible for creating the contentpanel for the Create Rule
  * XML Frame.
  */
-@Deprecated
+@Deprecated // to be removed with PMD 7.0.0
 public class CreateXMLRulePanel extends JPanel implements ActionListener {
 
     private JTextField rulenameField = new JTextField(30);