+++ pmd-core/src/main/java/net/sourceforge/pmd/util/designer/Designer.java
@@ -124,7 +124,7 @@ import net.sourceforge.pmd.lang.symboltable.Scope;
 import net.sourceforge.pmd.lang.symboltable.ScopedNode;
 import net.sourceforge.pmd.lang.xpath.Initializer;
 
-@Deprecated
+@Deprecated // to be removed with PMD 7.0.0
 public class Designer implements ClipboardOwner {
 
     private boolean exitOnClose = true;