+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/ActionCommands.java
@@ -9,7 +9,7 @@ package net.sourceforge.pmd.util.viewer.gui;
  *
  * @author Boris Gruschko ( boris at gruschko.org )
  */
-
+@Deprecated // to be removed with PMD 7.0.0
 public final class ActionCommands {
     public static final String COMPILE_ACTION = "Compile";
     public static final String EVALUATE_ACTION = "Evaluate";