+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/EvaluationResultsPanel.java
@@ -25,6 +25,7 @@ import net.sourceforge.pmd.util.viewer.model.ViewerModelListener;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
+@Deprecated // to be removed with PMD 7.0.0
 public class EvaluationResultsPanel extends JPanel implements ViewerModelListener {
     private ViewerModel model;
     private JList list;