+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/ParseExceptionHandler.java
@@ -25,7 +25,7 @@ import net.sourceforge.pmd.util.viewer.util.NLS;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
-
+@Deprecated // to be removed with PMD 7.0.0
 public class ParseExceptionHandler extends JDialog implements ActionListener {
     private Exception exc;
     private JButton okBtn;