+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/XPathPanel.java
@@ -23,7 +23,7 @@ import net.sourceforge.pmd.util.viewer.util.NLS;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
-
+@Deprecated // to be removed with PMD 7.0.0
 public class XPathPanel extends JTabbedPane implements ViewerModelListener {
     private ViewerModel model;
     private JTextArea xPathArea;