+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/menu/SimpleNodeSubMenu.java
@@ -18,6 +18,7 @@ import net.sourceforge.pmd.util.viewer.util.NLS;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
+@Deprecated // to be removed with PMD 7.0.0
 public class SimpleNodeSubMenu extends JMenu {
     private ViewerModel model;
     private Node node;