+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/gui/menu/XPathFragmentAddingItem.java
@@ -17,6 +17,7 @@ import net.sourceforge.pmd.util.viewer.model.ViewerModel;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
+@Deprecated // to be removed with PMD 7.0.0
 public class XPathFragmentAddingItem extends JMenuItem implements ActionListener {
     private ViewerModel model;
     private String fragment;