+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/model/ASTModel.java
@@ -20,7 +20,7 @@ import net.sourceforge.pmd.lang.ast.Node;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
-
+@Deprecated // to be removed with PMD 7.0.0
 public class ASTModel implements TreeModel {
 
     private Node root;