+++ pmd-core/src/main/java/net/sourceforge/pmd/util/viewer/model/SimpleNodeTreeNodeAdapter.java
@@ -19,7 +19,7 @@ import net.sourceforge.pmd.lang.ast.Node;
  * @author Boris Gruschko ( boris at gruschko.org )
  * @version $Id$
  */
-
+@Deprecated // to be removed with PMD 7.0.0
 public class SimpleNodeTreeNodeAdapter implements TreeNode {
 
     private Node node;